[refactor] ensure next button is disabled for invalid file type in upload jobdescription card
